requirements.txt
flask
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-urllib
opentelemetry-instrumentation-requests
opentelemetry-exporter-jaeger